pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crisai"
version = "0.1.0"
description = "Local AI tool platform for architecture and design work"
requires-python = ">=3.10"

dependencies = [
  "openai-agents>=0.2.10",
  "mcp>=1.10.0",
  "python-dotenv>=1.0.1",
  "PyYAML>=6.0.2",
  "typer>=0.12.3",
  "rich>=13.7.1"
]

[project.scripts]
crisai = "crisai.cli.main:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
